requirements.txt
gemini-engine>=1.0.7